cloud_server/requirements.txt
paho-mqtt==1.6.1
scikit-learn==1.6.1
numpy
xgboost